added functionality to favourite appearing

+++ src/app/shared/services/save-favourite-rocket.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import {BehaviorSubject, Subject} from 'rxjs';
+import {BehaviorSubject} from 'rxjs';
 
 @Injectable({
     providedIn: 'root'
@@ -8,23 +8,38 @@ export class SaveFavouriteRocketService {
   private rocket = new BehaviorSubject(JSON.parse(localStorage.getItem('favourite')));
   sharedRocket$ = this.rocket.asObservable();
   storageArr = [];
+  isExistRocketInStorage: boolean;
   constructor() { }
 
   saveToLocalStorage(data) {
     if (localStorage.getItem('favourite')) {
       this.storageArr = JSON.parse(localStorage.getItem('favourite'));
-      this.storageArr.push(data);
-      localStorage.setItem('favourite', JSON.stringify(this.storageArr));
-      this.nextRocket(data);
+      this.isExistedRocket(this.storageArr, data);
+      if (this.isExistRocketInStorage) {
+        return;
+      } else {
+        this.storageArr.push(data);
+        localStorage.setItem('favourite', JSON.stringify(this.storageArr));
+        this.addNextRocket(this.storageArr);
+      }
     } else {
       localStorage.setItem('favourite', JSON.stringify(this.storageArr));
+      this.isExistedRocket(this.storageArr, data);
       this.storageArr.push(data);
       localStorage.setItem('favourite', JSON.stringify(this.storageArr));
-      this.nextRocket(data);
+      this.addNextRocket(this.storageArr);
     }
   }
 
-  nextRocket(rocket) {
+  private isExistedRocket(storage, obj) {
+    console.log(storage, obj);
+    storage.forEach(el => {
+      this.isExistRocketInStorage =  el.id === obj.id;
+    });
+    console.log(this.isExistRocketInStorage);
+  }
+
+  addNextRocket(rocket) {
     this.rocket.next(rocket);
   }
 }
